telemetry/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pytz==2024.2
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6

pydantic-settings~=2.5.2
python-dotenv~=1.0.1
aiokafka~=0.11.0